requirements.txt
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
open_clip_torch==2.20.0
pyyaml==6.0.1
opencv-python-headless==4.8.1.78
tqdm==4.66.1
numpy
